feat: continue Ship class construction
write additional TDD tests in ship.test.js and
update the Ship class in ship.js to make the tests
pass.

add a validation check to the Ship class
constructor to ensure the length argument entered
is a number within the assigned range.

add new properties hitCount and isSunkStatus and
new methods hit() that keeps track of the number
of hits the ship has taken and isSunk() that
checks if the ship is sunk by comparing hitCount
and the length of the ship.