refactor: modify Gameboard logic and tests
add new test case for the receiveAttack method in
gameboard.test.js that takes coordinates to check
for ship hits.

build simple receiveAttack method in gameboard.js
that accepts coordinates and returns a message to
make the test pass as an initial step.

refactor Gameboard class properties and the
placeShip method in gameboard.js to refine the
logic and method of ship placement on the
gameboard, specifically, have ships cover the
correct number of coordinates based on their
starting coordinates, orientation, and length.