feat: add tests and functionality to receiveAttack
add additional receiveAttack tests to
gameboard.test.js to ensure hits and misses
register and are recorded, and create a mock to
test that a Ship's hit method is called on the
correct ship.

add shipObjects properties to the Gameboard class
in gameboard.js to allow newShip to save to when
the placeShip method is called.

add missCoordinates property to the Gameboard
class in gameboard.js to collect missed shot
coordinates.

build out receiveAttack method to search through
the arrays of each created ship to determine if
the attack coordinates received would hit or miss
any of the ships and then either have the hit Ship
run its hit method or record a miss to the
missCoordinates property.

rename some properties and variables for clarity.