feat: add tests for reportAllSunk method
add test to receiveAttack in gameboard.test.js
that has it report if a ship is sunk.

modify receiveAttack in gameboard.js to call the
Ship's isSunk method after a hit registers and if
it is sunk, add 1 to a new sunkShipCount property
in the Gameboard class.

create tests in gameboard.test.js for a new
reportAllSunk method in gameboard.js, the tests
check if Gameboard will report whether or not all
of its ships have been sunk.