feat: improve clicked cell eventListener logic
turn the eventListener in event.js into a function
called handleBoardClick.

add check to handleBoardClick to verify that if a
click causes a ship to sink it is logged to
console, and afterwards run reportAllSunk() to
check if the game is over and log to console if
true along with deactivating the eventListener so
additional clicks will no longer register on
either gameboard.

add check for class type in handleBoardClick to
verify if a cell has already been clicked to
prevent multiple clicks on the same cell, and if
not, assign it the 'attacked' class when clicked.