feat: add turn functionality
refactor startGame into into an encapsulating
function, housing new function alternateTurns and
the existing function handleAttackClick.

update the game text when startGame is run, when
turns alternate, and when a game over occurs.

modify handleAttackClick function in startGame in
events.js since it now has access to variables and
objects directly from startGame, also prevent
attack clicks on a player's own board.

create alternateTurns function in events.js that
switches which player is able to attack the
opponent's gameboard in an alternating sequence.