feat: add gameboard grid axis labels
import new typewriter style font to <head> in
template.html.

update createGrid in render.js to add axis labels
to the game grid by adding an extra row and column
for the labels. Labels are A-J for columns and
1-10 for rows. Adjust grid cell creation to
account for the extra cells used for labels.

add new styles for the axis labels in style.css.