feat: add pvc computer attacks
create handleComputerAttack function in
comp-logic.js that allows the computer to make
random attacks, but not in the same cell more than
once.

create wasCellAttacked function in the Gameboard
class in gameboard.js to assist
handleComputerAttack in checking if a cell was
previously targeted.

add a new property hitCoordinates to pair with the
existing missCoordinates in the Gameboard class to
assist wasCellAttacked with checks.